backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
